Rename LoanTemplate to Loan; implement getTotalInterest
Rename LoanTemplate.java to Loan.java and update the class and constructor names accordingly. Minor Javadoc/formatting tweaks (setLoanAmount comment). Implement getTotalInterest() to return getTotalPayment() - loanAmount (replacing the previous TODO). No other logic changes.
